Fixed test setups to pass on all Django versions
