+++ moderation/models.py
@@ -7,7 +7,7 @@ from django.db import models
 if VERSION >= (1, 8):
     from django.contrib.contenttypes.fields import GenericForeignKey
 else:
-    from django.contrib.contenttypes import GenericForeignKey
+    from django.contrib.contenttypes.generic import GenericForeignKey
 
 from moderation.diff import get_changes_between_models
 from moderation.fields import SerializedObjectField
@@ -42,7 +42,7 @@ class ModeratedObject(models.Model):
     object_pk = models.PositiveIntegerField(null=True, blank=True,
                                             editable=False)
     content_object = GenericForeignKey(ct_field="content_type",
-                                               fk_field="object_pk")
+                                       fk_field="object_pk")
     date_created = models.DateTimeField(auto_now_add=True, editable=False)
     date_updated = models.DateTimeField(auto_now=True)
     moderation_state = models.SmallIntegerField(choices=MODERATION_STATES,
