+++ moderation/utils.py
@@ -6,4 +6,3 @@ def django_17():
     if StrictVersion(django.get_version()) >= StrictVersion('1.7.0'):
         return True
     return False
-
