+++ tests/moderator.py
@@ -2,4 +2,5 @@ from __future__ import unicode_literals
 from moderation import moderation
 from .models import Book
 
+
 moderation.register(Book)
