+++ tests/more_models.py
@@ -1,0 +1,6 @@
+from django.db import models
+from moderation.db import ModeratedModel
+
+
+class MyTestModel(ModeratedModel):
+    name = models.CharField(max_length=20)
