+++ tests/tests/__init__.py
@@ -1,10 +1,10 @@
-from .unit.register import *
-from .unit.managers import *
-from .unit.models import *
-from .unit.admin import *
-from .unit.diff import *
-from .unit.forms import *
-from .unit.moderator import *
+from .unit.testregister import *
+from .unit.testmanagers import *
+from .unit.testmodels import *
+from .unit.testadmin import *
+from .unit.testdiff import *
+from .unit.testforms import *
+from .unit.testmoderator import *
 from .regression import *
-from .acceptance.exclude import *
-from .acceptance.auto_discover import *
+from .acceptance.testexclude import *
+from .acceptance.testauto_discover import *
