+++ tests/tests/unit/testmanagers.py
