+++ tox.ini
@@ -7,7 +7,8 @@
 envlist = django1.4-py26, django1.4-py27,
           django1.5-py26, django1.5-py27, django1.5-py33, django1.5-py34,
           django1.6-py26, django1.6-py27, django1.6-py33, django1.6-py34,
-            django1.7-py27, django1.7-py33, django1.7-py34
+          django1.7-py27, django1.7-py33, django1.7-py34, django1.8-py27,
+		  django1.8-py34
 
 [testenv]
 commands = python setup.py test
@@ -104,5 +105,19 @@ deps =
 basepython = python3.4
 deps =
     Django==1.7.0
+    unittest2py3k
+    {[testenv]deps}
+
+[testenv:django1.8-py27]
+basepython = python2.7
+deps =
+    Django==1.8.0
+    unittest2
+    {[testenv]deps}
+
+[testenv:django1.8-py34]
+basepython = python3.4
+deps =
+    Django==1.8.0
     unittest2py3k
     {[testenv]deps}
